+++ src/models/aimf.h
@@ -10,28 +10,28 @@
 #define MOD_AIMF 			23
 #define MOD_AIMFT 			24
 
-double AIMF_Dipolar_R1(double omega_obs, double omega_neigh, double taus, double S2s, double tauf, double S2f, double D);
-double AIMF_CSA_R2(double omega, \
-						 double w1, \
-						 double wr, \
-						 double taus, \
-						 double S2s, \
-						 double tauf, \
-						 double S2f, \
-						 double D2, \
-						 double  (*J_SD)(\
-						 	double,\
-						 	double, \
-						 	double, \
-						 	double, \
-						 	double)\
+Decimal AIMF_Dipolar_R1(Decimal omega_obs, Decimal omega_neigh, Decimal taus, Decimal S2s, Decimal tauf, Decimal S2f, Decimal D);
+Decimal AIMF_CSA_R2(Decimal omega, \
+						 Decimal w1, \
+						 Decimal wr, \
+						 Decimal taus, \
+						 Decimal S2s, \
+						 Decimal tauf, \
+						 Decimal S2f, \
+						 Decimal D2, \
+						 Decimal  (*J_SD)(\
+						 	Decimal,\
+						 	Decimal, \
+						 	Decimal, \
+						 	Decimal, \
+						 	Decimal)\
 						);
-double AIMF_Dipolar_R2(double omega_obs, double omega_neigh, double w1, double wr, double taus, double S2s, double tauf, double S2f, double D);
-int AIMF_S2(double order_params[3], struct Orient ** A, double * S2[], int length);
-double AIMF_15NR1(struct Residue *res, struct Relaxation* relax, double taus, double tauf, double * sigs, double * sigf);
-double AIMF_15NR2(struct Residue *res, struct Relaxation* relax, double taus, double tauf, double * sigs, double * sigf);
-double AIMF_13CR1(struct Residue *res, struct Relaxation* relax, double taus, double tauf, double * sigs, double * sigf);
-double AIMF_13CR2(struct Residue *res, struct Relaxation* relax, double taus, double tauf, double * sigs, double * sigf);
+Decimal AIMF_Dipolar_R2(Decimal omega_obs, Decimal omega_neigh, Decimal w1, Decimal wr, Decimal taus, Decimal S2s, Decimal tauf, Decimal S2f, Decimal D);
+int AIMF_S2(Decimal order_params[3], struct Orient ** A, Decimal * S2[], int length);
+Decimal AIMF_15NR1(struct Residue *res, struct Relaxation* relax, Decimal taus, Decimal tauf, Decimal * sigs, Decimal * sigf);
+Decimal AIMF_15NR2(struct Residue *res, struct Relaxation* relax, Decimal taus, Decimal tauf, Decimal * sigs, Decimal * sigf);
+Decimal AIMF_13CR1(struct Residue *res, struct Relaxation* relax, Decimal taus, Decimal tauf, Decimal * sigs, Decimal * sigf);
+Decimal AIMF_13CR2(struct Residue *res, struct Relaxation* relax, Decimal taus, Decimal tauf, Decimal * sigs, Decimal * sigf);
 
 #include "aimf.c"
 #endif
